requirements.txt
anthropic>=0.25.0
deepeval>=0.21.0
ollama>=0.2.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-html>=4.1.0
python-dotenv>=1.0.0
